+++ Samples/iOS-Swift/iOS-Swift/ExtraViewController.swift
@@ -2,7 +2,7 @@
 import AuthenticationServices
 import Foundation
 import SafariServices
-import Sentry
+@_spi(Private) import Sentry
 import SentrySampleShared
 import UIKit
 
@@ -212,25 +212,6 @@ class ExtraViewController: UIViewController {
         let errorEventID = SentrySDK.capture(error: NSError(domain: "test-error.user-feedback.iOS-Swift", code: 1))
         let feedback = SentryFeedback(message: "It broke again on iOS-Swift. I don't know why, but this happens.", name: "John Me", email: "john@me.com", source: .custom, associatedEventId: errorEventID, attachments: attachments)
         SentrySDK.capture(feedback: feedback)
-    }
-    
-    @IBAction func captureUserFeedback(_ sender: UIButton) {
-        highlightButton(sender)
-        let error = NSError(domain: "UserFeedbackErrorDomain", code: 0, userInfo: [NSLocalizedDescriptionKey: "This never happens."])
-
-        let eventId = SentrySDK.capture(error: error) { scope in
-            scope.setLevel(.fatal)
-        }
-
-      #if SDK_V9
-        print("SDK V9 does not support user feedback.")
-      #else
-        let userFeedback = UserFeedback(eventId: eventId)
-        userFeedback.comments = "It broke on iOS-Swift. I don't know why, but this happens."
-        userFeedback.email = "john@me.com"
-        userFeedback.name = "John Me"
-        SentrySDK.capture(userFeedback: userFeedback)
-      #endif // SDK_V9
     }
 
     @IBAction func permissions(_ sender: UIButton) {