+++ Sources/Sentry/_SentryDispatchQueueWrapperInternal.m
@@ -1,4 +1,5 @@
 #import "_SentryDispatchQueueWrapperInternal.h"
+#import "SentryInternalDefines.h"
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -23,16 +24,19 @@ NS_ASSUME_NONNULL_BEGIN
     return self;
 }
 
-- (void)dispatchAsyncWithBlock:(void (^)(void))block
+- (instancetype)initWithName:(const char *)name relativePriority:(int)relativePriority
 {
-    dispatch_async(_queue, ^{
-        @autoreleasepool {
-            block();
-        }
-    });
+    if (self = [super init]) {
+        SENTRY_CASSERT(relativePriority <= 0 && relativePriority >= QOS_MIN_RELATIVE_PRIORITY,
+            @"Relative priority must be between 0 and %d", QOS_MIN_RELATIVE_PRIORITY);
+        dispatch_queue_attr_t attributes = dispatch_queue_attr_make_with_qos_class(
+            DISPATCH_QUEUE_SERIAL, QOS_CLASS_UTILITY, relativePriority);
+        _queue = dispatch_queue_create(name, attributes);
+    }
+    return self;
 }
 
-- (void)dispatchAsyncOnMainQueue:(void (^)(void))block
+- (void)dispatchAsyncOnMainQueueIfNotMainThread:(void (^)(void))block
 {
     if ([NSThread isMainThread]) {
         block();
@@ -43,6 +47,15 @@ NS_ASSUME_NONNULL_BEGIN
             }
         });
     }
+}
+
+- (void)dispatchAsyncWithBlock:(void (^)(void))block
+{
+    dispatch_async(_queue, ^{
+        @autoreleasepool {
+            block();
+        }
+    });
 }
 
 - (void)dispatchSync:(void (^)(void))block